+++ apps/remnant2toolkit/app/(builds)/_libs/build-filters/limit-to-non-variant-builds.ts
@@ -1,0 +1,10 @@
+import { Prisma } from '@repo/db';
+
+export function limitToNonVariantBuilds(limitToNonVariantBuilds: boolean) {
+  if (!limitToNonVariantBuilds) return Prisma.empty;
+  return Prisma.sql`AND NOT EXISTS (
+  SELECT 1
+  FROM BuildVariant
+  WHERE BuildVariant.secondaryBuildId = Build.id
+)`;
+}