+++ apps/remnant2toolkit/app/(builds)/builder/[buildId]/_components/view-build.tsx
@@ -102,11 +102,6 @@ export function ViewBuild({
     });
   }
 
-  // Need to convert the build data to a format that the BuildPage component can use
-  if (!session?.user) {
-    mainBuildState.upvoted = false;
-  }
-
   // We need to convert the build.items object into an array of items to pass to the ToCsvButton
   const csvBuildData = buildStateToCsvData(activeBuildState);