+++ apps/remnant2toolkit/app/_libs/moderation/_types/discord-webhook-params.ts
@@ -1,0 +1,10 @@
+export type DiscordWebhookParams = {
+  embeds: Array<{
+    title: string;
+    color: number;
+    fields: Array<{
+      name: string;
+      value: string;
+    }>;
+  }>;
+};